Introduce the concept of zero bank to prevent emulator from crashing on out-of-bounds memory accesses.
MMU detects if address is not in the RAM/MMIO range and returns the pointer to an 8 bytes long memory area that
was not initialized with anything (No guarantees that it will contain all zeroes). Behaviour would be similar on
real hardware, I think.

This hack/workaround allows us to get further in the SMP kernel boot process -- we now reach kernel cmdline parsing.

Still haven't researched what the hell exactly it was doing to read out of bounds. Something related to IPIs
(Inter-processor interrupts)?